# Optimize `ImmutableMatrix`