There was quadratic complexity growth in `ImmutableDenseMatrix`
because `list(self.args[2])` was called every time `self._mat`
was accessed.  This patch passes the request directly through.

Tests show a small speedup for small matrices and a large speedup for big matrices.

Running

```
X = DeferredVector("X")
for i in range(1,11):
    w = 100*i
    mat = Matrix(1, w, lambda i, j: X[i*j])
    print(w)
    %timeit lambdify(X, mat)
```

produces 

```
100
100 loops, best of 3: 4.73 ms per loop
200
100 loops, best of 3: 9.19 ms per loop
300
100 loops, best of 3: 13.9 ms per loop
400
100 loops, best of 3: 18.9 ms per loop
500
10 loops, best of 3: 24.2 ms per loop
600
10 loops, best of 3: 29.3 ms per loop
700
10 loops, best of 3: 34.7 ms per loop
800
10 loops, best of 3: 40.6 ms per loop
900
10 loops, best of 3: 46.4 ms per loop
1000
10 loops, best of 3: 53 ms per loop
```

on master and

```
100
100 loops, best of 3: 4.55 ms per loop
200
100 loops, best of 3: 8.34 ms per loop
300
100 loops, best of 3: 12.3 ms per loop
400
100 loops, best of 3: 16.2 ms per loop
500
10 loops, best of 3: 20 ms per loop
600
10 loops, best of 3: 23.9 ms per loop
700
10 loops, best of 3: 27.8 ms per loop
800
10 loops, best of 3: 31.7 ms per loop
900
10 loops, best of 3: 35.7 ms per loop
1000
10 loops, best of 3: 39.9 ms per loop
```

with this PR.

With a matrix of size 10000, the difference is `396 ms` vs `1.55 s`