@siefkenj 2017-06-10
This partially addresses #12731 but doesn't get it fully up to the original speed.