@isuruf 2017-06-11
Here's a patch that further increases the speed.
```diff
diff --git a/sympy/matrices/common.py b/sympy/matrices/common.py
index b9c68e1..21b3dcf 100644
--- a/sympy/matrices/common.py
+++ b/sympy/matrices/common.py
@@ -103,11 +103,9 @@ def entry(i, j):
                                          lambda i, j: entry(i, j))
 
     def _eval_extract(self, rowsList, colsList):
-        mat = list(self)
-        cols = self.cols
-        indices = (i * cols + j for i in rowsList for j in colsList)
-        return self._new(len(rowsList), len(colsList),
-                         list(mat[i] for i in indices))
+        def entry(i, j):
+            return self[rowsList[i], colsList[j]]
+        return self._new(len(rowsList), len(colsList), entry)
 
     def _eval_get_diag_blocks(self):
         sub_blocks = []
```