@siefkenj 2017-06-12
@isuruf That patch actually slows things down significantly.  I got

```
100
100 loops, best of 3: 6.51 ms per loop
200
100 loops, best of 3: 12.6 ms per loop
300
100 loops, best of 3: 18.5 ms per loop
400
10 loops, best of 3: 24.3 ms per loop
500
10 loops, best of 3: 30.5 ms per loop
600
10 loops, best of 3: 36.5 ms per loop
700
10 loops, best of 3: 43.1 ms per loop
800
10 loops, best of 3: 48.7 ms per loop
900
10 loops, best of 3: 54.6 ms per loop
1000
10 loops, best of 3: 60.2 ms per loop
```

with the above benchmark.