@siefkenj 2017-06-12
Thanks to @isuruf 's suggestion, I optimized `_eval_extract` and got another 15% speed increase.

```
100
100 loops, best of 3: 3.77 ms per loop
200
100 loops, best of 3: 7.04 ms per loop
300
100 loops, best of 3: 10.3 ms per loop
400
100 loops, best of 3: 13.5 ms per loop
500
100 loops, best of 3: 16.9 ms per loop
600
10 loops, best of 3: 19.9 ms per loop
700
10 loops, best of 3: 23.5 ms per loop
800
10 loops, best of 3: 26.7 ms per loop
900
10 loops, best of 3: 30.1 ms per loop
1000
10 loops, best of 3: 33.3 ms per loop
```