@siefkenj 2017-06-13
@jksuom Could you review this?